axf/requirement.txt
django==2.1.7
pymysql==0.9.3
djangorestframework==3.9.2
django-filter==2.1.0
redis==3.2.1
django-redis==4.10.0
django-cors-headers
